Move calender settings into their own package
