+++ src/main/java/org/synyx/urlaubsverwaltung/calendarintegration/ExchangeCalendarSettings.java
@@ -1,4 +1,4 @@
-package org.synyx.urlaubsverwaltung.settings;
+package org.synyx.urlaubsverwaltung.calendarintegration;
 
 import javax.persistence.Column;
 import javax.persistence.Embeddable;
