+++ src/main/java/org/synyx/urlaubsverwaltung/calendarintegration/providers/noop/NoopCalendarSyncProvider.java
@@ -4,7 +4,7 @@ import org.slf4j.Logger;
 import org.springframework.stereotype.Service;
 import org.synyx.urlaubsverwaltung.absence.Absence;
 import org.synyx.urlaubsverwaltung.calendarintegration.providers.CalendarProvider;
-import org.synyx.urlaubsverwaltung.settings.CalendarSettings;
+import org.synyx.urlaubsverwaltung.calendarintegration.CalendarSettings;
 
 import java.util.Optional;
