+++ src/test/java/org/synyx/urlaubsverwaltung/calendarintegration/ExchangeCalendarSettingsTest.java
@@ -1,6 +1,7 @@
-package org.synyx.urlaubsverwaltung.settings;
+package org.synyx.urlaubsverwaltung.calendarintegration;
 
 import org.junit.jupiter.api.Test;
+import org.synyx.urlaubsverwaltung.calendarintegration.ExchangeCalendarSettings;
 
 import static org.assertj.core.api.Assertions.assertThat;
